libft

+++ libft.h
@@ -36,7 +36,7 @@ char	*ft_strdup(char	*src);
 void	ft_striteri(char *s, void (*f)(unsigned int, char*));
 char	*ft_strcat(char	*dest, char	*src);
 char 	*ft_strjoin(char const *s1, char const *s2);
-unsigned int	ft_strlcat(char	*  dest, char	* src, size_t	dstsize);
+size_t	ft_strlcat(char	*  dest, char	* src, size_t	dstsize);
 size_t	ft_strlcpy(char	* dest, char	* src, size_t	dstsize);
 int		ft_strlen(const char	*str);
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char));
